update routes/controllers (moved the character search from the characters controller to the api controller)
