+++ public/css/styles.css
@@ -37,7 +37,7 @@ nav {
   background-color: #B6050F;
   position: fixed;
   top: 0;
-  z-index: 10000; }
+  z-index: 500; }
   nav ul {
     list-style-type: none;
     margin: 0;
