+++ public/css/styles.scss
@@ -76,7 +76,7 @@ nav {
   background-color: $red;
   position: fixed;
   top: 0;
-  z-index: 10000;
+  z-index: 500;
 
   ul {
     list-style-type: none;
